# api/internal/server/assets/openapi.yaml
openapi: 3.1.0
info:
  title: adVideo — API de Integrações
  version: "1.0.0"
  summary: Download de mídia de várias plataformas, consumido por outro sistema.
  description: |
    Esta é a API que **outros sistemas** usam para pedir downloads à plataforma.
    É a mesma engrenagem que atende a interface web — mesma fila, mesmos
    provedores, mesmo armazenamento —, com autenticação por chave de API em vez
    de login, e com cota, limites e auditoria por integração.

    ## Como funciona, em uma frase

    Você faz `POST /v1/integration/downloads` com uma URL, recebe `202` com um
    identificador, e nós avisamos o seu sistema por **webhook** quando o arquivo
    fica pronto. Enquanto isso, `GET /v1/integration/downloads/{id}` responde o
    estado atual a qualquer momento.

    ## Autenticação

    Toda chamada exige a chave de API emitida no painel:

    ```
    Authorization: Bearer adk_live_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    Se o seu ambiente reescreve o header `Authorization` (proxies corporativos
    às vezes fazem isso), use `X-API-Key` com o mesmo valor.

    Opcionalmente, envie também `X-Integration-Id` com o identificador público
    da integração. Ele não substitui a chave: serve como trava contra o acidente
    de apontar a chave de homologação para a integração de produção — se os dois
    não corresponderem, a chamada é recusada com `integration_mismatch`.

    **A chave aparece uma única vez**, na resposta que a criou. O banco guarda
...

...

...

...

...
servers:
...
security:
  - ChaveDeAPI: []
...
paths:
...
  /v1/integration/me:
    get:
      tags: [Integração]
      summary: Quem esta chave representa
      description: |
        Primeira chamada de qualquer integração nova: confirma que a credencial
        funciona, devolve o identificador público e mostra os limites em vigor.
      operationId: obterIntegracao
      responses:
        "200":
          description: Dados da integração autenticada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  integration:
                    $ref: "#/components/schemas/Integration"
                  limits:
                    $ref: "#/components/schemas/Limits"
                  usage:
                    $ref: "#/components/schemas/Quota"
              example:
                integration:
                  id: 9f1c7a32-4b8e-4a1d-9c3f-2e5b7d8a1f04
                  name: CRM Comercial
                  active: true
                  created_at: "2026-09-01T13:04:11Z"
                limits:
                  daily_downloads: 200
                  max_concurrent_downloads: 5
                  rate_limit_per_minute: 120
                  max_file_size_bytes: 1073741824
                  allowed_ips: ["203.0.113.10", "198.51.100.0/24"]
                usage:
                  daily_limit: 200
                  daily_used: 37
                  daily_remaining: 163
                  unlimited: false
                  downloads_active: 2
                  concurrency_limit: 5
                  resets_at: "2026-09-13T00:00:00-03:00"
                  resets_at_timezone: America/Sao_Paulo
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "403": { $ref: "#/components/responses/Proibido" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
  /v1/integration/quota:
    get:
      tags: [Integração]
      summary: Consumo da cota diária
      description: |
        Resposta enxuta para consultar antes de disparar um lote. Chamar esta
        rota não consome cota (consome apenas do limite por minuto).
      operationId: obterCota
      responses:
        "200":
          description: Consumo atual.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quota" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
  /v1/integration/media/resolve:
    post:
      tags: [Mídia]
      summary: Consultar metadados e formatos de uma URL
      description: |
        Identifica a plataforma e devolve título, duração, capa e a lista de
        formatos disponíveis. **Não cria download e não consome cota diária.**

        Use esta rota quando precisar escolher a qualidade: o `id` de cada
        formato é o valor aceito em `format_id` na criação do download.

        Os identificadores de formato **não são estáveis** entre duas consultas
        do mesmo conteúdo — a plataforma os regenera. Resolva e crie o download
        na mesma sequência; se o identificador tiver expirado no meio do
        caminho, a criação cai automaticamente para a melhor qualidade
        equivalente em vez de falhar.
      operationId: resolverMidia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  description: URL pública do conteúdo.
                  examples: ["https://www.youtube.com/watch?v=dQw4w9WgXcQ"]
      responses:
        "200":
          description: Metadados do conteúdo.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaMetadata" }
        "400":
          description: URL inválida ou ausente.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Plataforma não suportada ou conteúdo inexistente.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: |
            Conteúdo privado, restrito por região, ao vivo, ou que exige login
            indisponível. Códigos: `content_private`, `geo_blocked`,
            `live_content`, `auth_required`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }
        "503":
          description: Mecanismo de download indisponível no momento.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # -------------------------------------------------------------------------
  /v1/integration/downloads:
    post:
      tags: [Downloads]
      summary: Criar um download
      description: |
        Aceita o pedido e devolve `202` com o download em `PENDING` ou
        `PROCESSING`. O arquivo **não existe ainda**: acompanhe por webhook ou
        consultando `GET /v1/integration/downloads/{id}`.

        Consome uma unidade da cota diária e um slot de simultaneidade.
      operationId: criarDownload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                  description: URL pública do conteúdo.
                kind:
                  type: string
                  enum: [video, audio]
                  default: video
                  description: |
                    `video` entrega MP4; `audio` extrai a trilha e entrega MP3.
                format_id:
                  type: string
                  description: |
                    Formato escolhido, vindo de `/media/resolve`. Omitido, usa a
                    melhor qualidade disponível do tipo pedido.
                  examples: ["137"]
            examples:
              melhorQualidade:
                summary: Vídeo na melhor qualidade
                value: { url: "https://www.youtube.com/watch?v=dQw4w9WgXcQ" }
              somenteAudio:
                summary: Só o áudio, em MP3
                value:
                  url: "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                  kind: audio
              qualidadeEscolhida:
                summary: Qualidade específica vinda de /media/resolve
                value:
                  url: "https://vimeo.com/76979871"
                  kind: video
                  format_id: "http-1080p"
      responses:
        "202":
          description: Pedido aceito e enfileirado.
          headers:
            X-Quota-Remaining:
              schema: { type: integer }
              description: Downloads restantes na cota de hoje.
            X-Quota-Reset:
              schema: { type: string, format: date-time }
              description: Quando a cota reseta.
          content:
            application/json:
              schema:
                type: object
                properties:
                  download: { $ref: "#/components/schemas/Download" }
        "400":
          description: |
            Pedido inválido, URL não aceita, ou arquivo acima do limite da
            integração. Códigos: `invalid_request`, `invalid_url`,
            `format_unavailable`, `file_too_large`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "403": { $ref: "#/components/responses/Proibido" }
        "404":
          description: Plataforma não suportada ou conteúdo inexistente.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Conteúdo indisponível para download (privado, ao vivo, restrito).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: |
            Um dos três limites. Distinga pelo `code`: `rate_limited` (chamadas
            por minuto), `quota_exceeded` (cota diária) ou `too_many_concurrent`
            (downloads em andamento).
          headers:
            Retry-After:
              schema: { type: integer }
              description: Segundos a aguardar. Ausente em `quota_exceeded`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                cota:
                  value: { error: "cota diária de downloads esgotada", code: quota_exceeded }
                simultaneos:
                  value:
                    error: "limite de downloads simultâneos atingido; aguarde a conclusão dos atuais"
                    code: too_many_concurrent
        "503":
          description: Mecanismo de download indisponível no momento.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    get:
      tags: [Downloads]
      summary: Listar downloads
      description: |
        Lista, do mais recente para o mais antigo, apenas os downloads desta
        integração.

        É também a rede de segurança do webhook: se o seu endpoint ficou fora do
        ar, `?status=COMPLETED` recupera o que você não recebeu.
      operationId: listarDownloads
      parameters:
        - { $ref: "#/components/parameters/Pagina" }
        - { $ref: "#/components/parameters/PorPagina" }
        - name: status
          in: query
          description: Filtra por situação. Maiúsculas, exatamente como no enum.
          schema: { $ref: "#/components/schemas/DownloadStatus" }
        - name: platform
          in: query
          description: Filtra por plataforma de origem.
          schema: { type: string, examples: ["youtube"] }
        - name: search
          in: query
          description: Busca no título e na URL original.
          schema: { type: string }
      responses:
        "200":
          description: Página de downloads.
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloads:
                    type: array
                    items: { $ref: "#/components/schemas/Download" }
                  pagination: { $ref: "#/components/schemas/Pagination" }
        "400":
          description: Filtro inválido (por exemplo, `status` fora do enum).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
  /v1/integration/downloads/{id}:
    parameters:
      - { $ref: "#/components/parameters/DownloadID" }

    get:
      tags: [Downloads]
      summary: Consultar um download
      description: |
        Estado atual do download. É a alternativa oficial ao webhook para quem
        não pode expor um endpoint público.

        Consultando em laço, respeite o limite por minuto: um intervalo de 5 a
        10 segundos por download em andamento é suficiente, já que o progresso
        muda nessa ordem de tempo.
      operationId: obterDownload
      responses:
        "200":
          description: O download.
          content:
            application/json:
              schema:
                type: object
                properties:
                  download: { $ref: "#/components/schemas/Download" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

    delete:
      tags: [Downloads]
      summary: Remover um download
      description: |
        Remove o download do histórico e apaga o arquivo do armazenamento. Não
        devolve a unidade de cota consumida — a cota conta pedidos feitos no
        dia, não arquivos guardados.
      operationId: removerDownload
      responses:
        "204": { description: Removido. }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
  /v1/integration/downloads/{id}/cancel:
    parameters:
      - { $ref: "#/components/parameters/DownloadID" }
    post:
      tags: [Downloads]
      summary: Cancelar um download
      description: |
        Interrompe um download em andamento. O processo que está baixando é
        avisado e para em poucos segundos.

        **É idempotente.** Cancelar algo que acabou de terminar sozinho não é
        erro: a resposta devolve o status que de fato vigora, e pode ser
        `COMPLETED`. Verifique o `status` da resposta em vez de assumir
        `CANCELED`.
      operationId: cancelarDownload
      responses:
        "200":
          description: Pedido registrado; o download traz o status vigente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  download: { $ref: "#/components/schemas/Download" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
  /v1/integration/downloads/{id}/download-url:
    parameters:
      - { $ref: "#/components/parameters/DownloadID" }
    get:
      tags: [Downloads]
      summary: Obter URL temporária do arquivo
      description: |
        Devolve uma URL assinada, **válida por 1 minuto**, apontando direto para
        o armazenamento. É a forma recomendada de buscar o arquivo: os bytes não
        passam pela API, então não há limite de tamanho nem timeout de
        requisição no caminho.

        Não guarde essa URL: ela expira. Guarde o `id` do download e peça uma
        nova quando precisar.
      operationId: obterURLDoArquivo
      responses:
        "200":
          description: URL assinada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, format: uri }
                  expires_at: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "409":
          description: O download ainda não terminou (`download_not_ready`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "410":
          description: O arquivo expirou e foi removido (`download_expired`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
  /v1/integration/downloads/{id}/file:
    parameters:
      - { $ref: "#/components/parameters/DownloadID" }
    get:
      tags: [Downloads]
      summary: Baixar o arquivo pela API
      description: |
        Entrega os bytes do arquivo pela própria API, com
        `Content-Disposition: attachment`.

        Prefira `/download-url` quando possível: aqui os bytes atravessam a API,
        o que consome banda do servidor e fica sujeito ao timeout do seu cliente
        HTTP em arquivos grandes. Esta rota existe para ambientes que não podem
        alcançar o armazenamento diretamente.
      operationId: baixarArquivo
      responses:
        "200":
          description: O arquivo.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "409":
          description: O download ainda não terminou (`download_not_ready`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "410":
          description: O arquivo expirou e foi removido (`download_expired`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
  /v1/integration/webhooks:
    get:
      tags: [Webhooks]
      summary: Listar webhooks configurados
      description: |
        Mostra os endpoints cadastrados, os eventos assinados e o diagnóstico da
        última entrega.

        O **segredo de assinatura não é devolvido aqui**, por decisão de
        segurança: ele é entregue pelo painel a uma pessoa. Se ele saísse por
        uma rota autenticada pela chave de API, um vazamento de chave viraria
        também um vazamento do segredo — e com ele seria possível forjar eventos
        assinados para o seu endpoint.

        O cadastro e a rotação são feitos no painel, não por esta API.
      operationId: listarWebhooks
      responses:
        "200":
          description: Webhooks da integração.
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items: { $ref: "#/components/schemas/Webhook" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
  /v1/integration/webhooks/{webhookId}/test:
    parameters:
      - name: webhookId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Webhooks]
      summary: Disparar evento de teste
      description: |
        Enfileira um evento `ping` para o endpoint cadastrado. Serve para
        validar a verificação de assinatura do seu lado sem esperar um download
        real.

        Responde `202` porque o POST no seu endpoint acontece em segundo plano;
        o resultado aparece em `GET /v1/integration/deliveries`.
      operationId: testarWebhook
      responses:
        "202":
          description: Teste enfileirado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "429": { $ref: "#/components/responses/LimiteExcedido" }

  # -------------------------------------------------------------------------
...
components:
  securitySchemes:
    ChaveDeAPI:
      type: http
      scheme: bearer
      description: |
        Chave de API emitida no painel, com o prefixo `adk_live_`.

        Alternativa para ambientes que reescrevem `Authorization`: o header
        `X-API-Key` com o mesmo valor.
  parameters:
    DownloadID:
      name: id
      in: path
      required: true
      description: Identificador do download.
      schema: { type: string, format: uuid }
    Pagina:
      name: page
      in: query
      description: Página, a partir de 1.
      schema: { type: integer, minimum: 1, default: 1 }
...